pyproject.toml
[project]
name = "Course_Plataform_-_Backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-polymorphic>=4.1.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "python-dotenv>=1.1.1",
    "drf-social-oauth2>=3.0.0",
    "social-auth-app-django>=5.0.0",
    "drf-spectacular>=0.27.0",
    "django-filter>=24.1",
    "gunicorn>=21.2.0",
    "psycopg2-binary>=2.9.9",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
]
